Refactor and cleanup DDlogJooqProvider
